@@ -25,30 +25,33 @@
   },
   "dependencies": {
     "antd": "^2.11.1",
+    "axios": "^0.15.3",
     "classnames": "^2.2.5",
+    "d3-shape": "^1.2.0",
+    "draft-js": "^0.10.1",
+    "draftjs-to-html": "^0.7.0",
     "draftjs-to-markdown": "^0.4.0",
     "dva": "^1.2.1",
     "dva-loading": "^0.2.0",
-    "lodash": "^4.17.4",
-    "nprogress": "^0.2.0",
-    "prop-types": "^15.5.10",
-    "qs": "^6.2.0",
-    "react": "^15.4.4",
-    "react-countup": "1.3.0",
-    "react-dom": "^15.4.1",
-    "react-draft-wysiwyg": "^1.8.1",
-    "react-helmet": "^5.0.0",
-    "recharts": "^0.19.0",
-    "axios": "^0.15.3",
-    "jsonp": "^0.2.1",
-    "path-to-regexp": "^1.7.0",
-    "rc-tween-one": "^1.0.0",
-    "redbox-react": "^1.3.2",
-    "draftjs-to-html": "^0.7.0",
     "dva-model-extend": "^0.1.1",
     "expect": "^1.20.2",
     "husky": "^0.12.0",
+    "jsonp": "^0.2.1",
+    "lodash": "^4.17.4",
     "mockjs": "^1.0.1-beta3",
+    "moment": "^2.18.1",
+    "nprogress": "^0.2.0",
+    "path-to-regexp": "^1.7.0",
+    "prop-types": "^15.5.10",
+    "qs": "^6.2.0",
+    "rc-tween-one": "^1.2.7",
+    "react": "^15.6.1",
+    "react-countup": "1.3.0",
+    "react-dom": "^15.4.1",
+    "react-draft-wysiwyg": "^1.10.0",
+    "react-helmet": "^5.0.0",
+    "recharts": "^0.19.0",
+    "redbox-react": "^1.3.2",
     "roadhog": "0.6.0-beta.6"
   },
   "devDependencies": {
