@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Table } from 'antd'
-import { request } from '../../utils'
 import lodash from 'lodash'
+import { request } from '../../utils'
 import './DataTable.less'
 
 class DataTable extends React.Component {
