@@ -2,8 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Breadcrumb, Icon } from 'antd'
 import { Link } from 'dva/router'
-import styles from './Bread.less'
 import pathToRegexp from 'path-to-regexp'
+import styles from './Bread.less'
 import { queryArray } from '../../utils'
 
 const Bread = ({ menu }) => {
