@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styles from './Container.less'
 import { ResponsiveContainer } from 'recharts'
+import styles from './Container.less'
 
 const Container = ({ children, ratio = 5 / 2, minHeight = 250, maxHeight = 350 }) => <div className={styles.container} style={{ minHeight, maxHeight }}>
   <div style={{ marginTop: `${100 / ratio}%` || '100%' }} />
