@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styles from './completed.less'
 import classnames from 'classnames'
-import { color } from '../../../utils'
 import { AreaChart, Area, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer } from 'recharts'
+import styles from './completed.less'
+import { color } from '../../../utils'
 
 function Completed ({ data }) {
   return (
