@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styles from './cpu.less'
-import { color } from '../../../utils'
 import CountUp from 'react-countup'
 import { LineChart, Line, XAxis, YAxis, CartesianGrid, ResponsiveContainer } from 'recharts'
+import styles from './cpu.less'
+import { color } from '../../../utils'
 
 const countUpProps = {
   start: 0,
