@@ -1,8 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import { Form, Button, Row, Col, DatePicker, Input, Cascader, Switch } from 'antd'
 import moment from 'moment'
 import { FilterItem } from '../../components'
-import { Form, Button, Row, Col, DatePicker, Input, Cascader, Switch } from 'antd'
+
 import city from '../../utils/city'
 
 const Search = Input.Search
