@@ -3,7 +3,7 @@ import { request, config } from '../utils'
 const { api } = config
 const { userLogin } = api
 
-export async function login (data) {
+export default async function login (data) {
   return request({
     url: userLogin,
     method: 'post',
