@@ -3,7 +3,7 @@ import { request, config } from '../utils'
 const { api } = config
 const { posts } = api
 
-export async function query (params) {
+export default async function query (params) {
   return request({
     url: posts,
     method: 'get',
